ENH: Use constructors with PropertyMaps in Control::SolverFactory
This is to make configuring solvers with PropertyMaps easier, and to
enable writing of a Solver's configuration to a PropertyMap, e.g. to
save this configuration for reproducibility.

Previously, Control::SolverFactory parsed the parameters from a
PropertyMap and then passed them to the constructor or
configure_iterative_solver(). This has been replaced by passing the
PropertyMap to the solver's constructor, or rather the
new_INSERT_SOLVER_NAME_HERE() routine calling the constructor. This in
turn means that almost all solvers received another constructor which
receives a String identifying the section in the PropertyMap used for
configuring it, as well as said section.

As a consequence, there are now two valid ways to create solver objects
through their respective new_INSERT_SOLVER_NAME_HERE() routines:
 1) Use the old functionality, meaning passing all relevant variables to
    the function. The resulting solver object will not know anything
    about how it was configured
 2) Use the new version with a PropertyMap. The resulting solver object
    will then know its orgins.

The base for this is in Solver::Base. The standard constructor works
like before and the new get_section_name() routine will just return the
solver's name(). However, if the new PropertyMap based constructor was
used, the object will know from which section it was configured.

Also got rid of the old BasicVCycle class, as it was no longer used
anywhere.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>