Removed MeshoptAssemblerLevel(s)
As the DomainLevel now has the transformation and FE spaces, it was
possible to move all functionality into the system levels. There is now
a base class called MeshoptSystemLevel which has almost all
functionality, from which QuadraticSystemLevel and NonlinearSystemLevel
are derived.

Also reordered and revised the visibility of the member variables in
DuDvFunctional and HyperelasticityFunctional.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>