BiCGStab rewrite
This implementation is directly derived from Van der Vorst's publication
and preconditioning can be done left-only or right-only. It does not
include any fancy (and/or heuristic) look-ahead or restart strategies.

The preconditioning defaults to right-only, because this only requires
two applications of the preconditioner. Superficial tests show that the
left-only variant may perform significantly better in terms of iteration
numbers, though.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>