Revised BiCGStab implementation yet again
The previous direct implementation of the pseudo code in Van der Vorst's
paper required one additional application of the preconditioner in the
left-only preconditioned variant. Peter unrolled the code again by hand
and was able to avoid this. The new implementation also needs fewer
temporary vectors and is reordered to be similar e.g. to the PCG
implementation.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>