Chebyshev smoother
Added Chebyshev solver class wich should ony be used as a smoother in
multigrid.
By default, it smoothes out eigenvalues between 0.5 and 0.8 of the
maximum eigenvalue of the given linear system.
This intervall can be changed with the parameters fraction_min_ev and
fraction_max_ev.
As a rule of thumb, you should use four or five iterations to get a
descent polynomial degree.

Signed-off-by: Dirk Ribbrock <dirk.ribbrock@math.tu-dortmund.de>