Added support for cuda 8
As cusolver depends on openmp in cuda 8, the -fopenmp flag is passed to
the host compilers g++ and icpc.
Until now, clang is not supported with cuda 8 due to its non conformal
openmp activation.

In addition, the default compute capability was raised to sm_30, as even
the nvcc compiler itself marks sm_20 as deprecated.

Signed-off-by: Dirk Ribbrock <dirk.ribbrock@math.tu-dortmund.de>