changed return value of Math::invert_matrix()
Previously, the Math::invert_matrix function returned the determinant of
the input matrix, which could be tested for matrix singularity via the
Math::isnormal function. Unfortunately, checking the determinant may
result in false alerts, as happened in the 3D Q2 Vanka smoother, as all
eigenvalues of the input matrix may be small, but the matrix may still
be well conditioned.

As a result, the Math::invert_matrix function now returns the quotient
of the maximum and minimum pivot elements, which occured during the
Gauss elimination process. Although this value usually does not
correspond to the condition number of the inverted matrix, it does at
least provide a value that is invariant to matrix scaling and it should
work as a failure/success indicator, too.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>