partially reverted previous commit
Turns out that some functions rely on the fact that Math::invert_matrix
in fact returns the determinant of the input matrix...

Therefore, the changes to this function from the previous commit have
been reverted. As compensation, the Vanka smoother does not throw an
exception when the returned determinant is subnormal or zero, but
instead ignores the return value.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>