Threaded mpi execution
This commits adds support for enforced asynchronous mpi execution.
The SynchScalarTicket spawns another thread, were iallreduce and wait
are executed without further delay.

To work correctly, your mpi implementation needs to support
MPI_THREAD_MULTIPLE, which can be checked with the mpi-info tool in the
tools folder.
Afterwards, the corresponding configure flag is --mpi-thread-multiple.

The synch_scal.hpp is hopefully not in its final stage, but may serve as
a good example on how to use #ifdef statements to obfuscate a methods
functionality in the mean time.

Signed-off-by: Dirk Ribbrock <dirk.ribbrock@math.tu-dortmund.de>