added MPI support for __float128 datatype
This commit adds MPI support for the 128-bit floating point
datatype "__float128" supported by GCC if FEAT is configured
with "quadmath" enabled.

To achieve this, some internal modifications had to be made:

* The Runtime class does not initialise MPI by itself anymore,
  but instead calls the new "Dist::initialise" function to do
  this. This new function also declares a corresponding MPI
  datatype handle for __float128, which is wrapped via
  "Dist::dt__float128".

* Also, the Runtime class now calls "Dist::finalise" to free
  all MPI-related resources as the new datatype.

Moreover, this commit also adds the '--ovr_mpi_ops' option
to the configure script, which defines "FEAT_OVERRIDE_MPI_OPS".
This define causes the Dist implementation to provide custom
MPI operation implementations for the "op_sum", "op_max" and
"op_min" wrappers. This was necessary as one cannot provide
custom implementations for the predefined MPI operations for
user-defined datatypes. These overrides are disabled if the
previously mentioned macro is not defined, as using custom
operation implementations seems to have a negative performance
impact.

In summary:
Unless you want to use the "__float128" in a parallel simulation,
nothing changes for you. If you want to use that, you will also
have to specify the '--ovr_mpi_ops' option to the configure
script, as otherwise MPI will abort at the first time when a
reduction operation is applied onto a "__float128" operand.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>