added Trafo::InverseMapping class template
The new InverseMapping class template can be used to "unmap" points
given in real world coordinates onto a set of mesh elements and their
corresponding reference cell coordinates, which map to the given input
point under the standard "forward" transformation.

This class template wraps around a Trafo::Standard::Mapping object and
provides the following basic functionality:

* Find a set of candidate cells that may intersect with a given point by
  a simple bounding-box test.

* Unmap a point given in world coordinates back onto the reference cell
  by using a Newton iteration.

This new class is a core component used by the upcoming DiscreteEvaluator
class template, which is responsible for evaluating a finite element
function in an arbitrary point.

A corresponding unit-test is included.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>