added Assembly::DiscreteEvaluator class
This new class provides several functions which can be used to evaluate
scalar or vector-valued finite element functions in points given in real
world coordinates. These functions rely on the Trafo::InverseMapping
class template perform the actual point unmapping.

Currently, the class only provides the functionality to evaluate
function values, but not gradients or hessians. This functionality can
and will be implemented later. Furthermore, this class may be extended
to evaluate other discrete (non-fe) functions defined on a mesh.

A corresponding unit-test is included.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>