added flow accumulator assembly to TraceAssembler
The TraceAssembler has been enhanced by a function which can assemble
certain values deriving from a velocity-pressure combination. This
function can be used to assemble drag & lift forces for various
benchmark problems.

Note that this function is somewhat rudimentary/provisional, but it does
the job. We may want to find a smarter solution later.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>