Iterative Partitioner
This commits adds a new iterative partitioner to the partitioner zoo.
The partitioner is based on randomly choosen initial patch centers.
The best guesses are then optimised iterativly by a randomised mutation
heuristic.

In the absence of a strong convergence criterium, the amount of time
spend in each phase can be controlled and is set to 3 seconds each by
default.

Signed-off-by: Dirk Ribbrock <dirk.ribbrock@math.tu-dortmund.de>