added support for virtual hierarchies to Multigrid
This commit adds support for virtual mesh hierarchies to the Multigrid
solver. The only significant change to the existing applications is that
one now needs to pass the virtual domain size to the constructor of the
multigrid hierarchy or matrix strock object.

Currently, "poisson_neumann_2d" is the only application making use of
truly virtual mesh hierarchies.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>