FIX: Switched icc from AVX2 to AVX-I for Broadwell and up
Haswell, Broadwell and Skylake do support -xcore-AVX2, but all Intel
compilers up to 16.0.4 produce code with significantly lower precision,
most notably in kernel/solver/optimiser-test. This is the case for both
Haswell and Skylake CPUs, so Broadwell will be affected as well.
At this point it is not clear if there is a compiler bug or this is
normal behaviour and AVX2 reduces the precision, so deactivate this for
the time being.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>