ENH: Moved operator splitting BDF(q) time disc to Control
Before, this was hard-coded into the screws application, where it COULD
have stayed ad infinitum as it is quite special code only for solving
the incompressible Navier-Stokes equations with a pressure projection
method. But now it is documented in Doxygen so it can be found and used
for other applications.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>