ENH: Mixed discretisation for Poisson
This application solves the Poisson problem in mixed formulation using
an inf-sup stable FE pair with discontinuous primal space (e.g.
discontinuous P1, which means the dual space is something like Q2). This
is usually nonsensical, but sometimes offers an advantage, e.g. if a
Q2/P1 discretisation of the Navier-Stokes equations is to be solved.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>