ENH: Added row_norm2() to SparseMatrixCSR/BCSR/ELL for Mem::Main
This computes the 2-norm of every row and saves it to an L-vector.
Useful for extracting the diagonal of a B^T diag(a) B matrix.

Note that this is not implemented for the Global::Matrix. The operation
would require a type 1 version of the matrix, which is expensive and is
currently only used for matrices where type 0 and type 1 are the same.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>