STYLE: Call it CellFunctional
There is the Global::NonlinearFunctional class, which is the global
wrapper around (patch-)local nonlinear functionals (of which the
Meshopt::HyperelasticityFunctional is the only member atm.). The
Meshopt::HyperelasticityFunctional has a (cell-)local functional
(similar to a linear operator's local (multi)linear form), which is now
properly called exactly that.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>