Reduced MPI ranks from 4 to 2 for MeshoptRefinementApp test
Running this test with MPI produces non-deterministic results due to the
indeterminate order of summation e.g. in dot products. As this problem
is both strongly nonlinear and "rough" (lots of if-else in the line
search) with the code branching into very different cases, this creates
a problem with the non-uniqueness of the solution. Meaning the achieved
minimum can be very different on two subsequent runs, which makes this a
poor test.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>