removed automagic in-place detection in Comm::gather/scatter
The automatic detection of whether to choose MPI_IN_PLACE for the
various variants of gather/scatter operations has been removed. In the
unlikely event that you really want to use MPI_IN_PLACE in one of those
ops, you'll have to do that by your own.

Note that this commit does not alter the behaviour of the reduction
operations, i.e. the [i][all]reduce() and [i|ex]scan() functions still
support the automatic detection of MPI_IN_PLACE in the case that the
send and receive buffers coincide.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>