Mirror, Mirror on the wall, you're the next to overhaul...
This commit provides a major rewrite of the VectorMirror class as well
as the other vector and matrix mirror classes. The VectorMirror class
now only contains an array of indices instead of two CSR matrices and
has been modified to derive from the LAFEM::Container base class
directly.

This rewrite has become necessary as part of the preparations for
EXA-scale computations, as the previous variant would have inevitably
become a show-stopper in combination with the Global::Muxer class...

Most of the algorithms have been rewritten for the new data structures,
merely the MatrixMirror operations for SparseMatrixELL and
SparseMatrixBanded are currently missing - this will be reimplemented
later.

On the bright side, the changes introduced to the mirror classes should
have a positive impact on runtime and memory usage, although it is
unclear whether the gain is significant.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>