implemented HierarchUnitCubeDomainControl2
This new experimental domain control is capable of creating recursively
partitinoned unit-square multi-level hierarchies. In contrast to the
older HierarchUnitCubeDomainControl (w/o 2) it can also create
hierarchies with 4^k:1 recursions instead of 4:1 recursion only.

The main purpose of this experimental domain control is to set up a
global multigrid on a truly recursive partitioning (e.g. 4096:64:1) on a
really large machine (i.e. >= 100K cores) as a "proof-of-concept".

As usual, don't use this unless you really know what you are doing.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>