use PartiDomainControl in poisson_neumann app
Until now, poisson_neumann also served as a testing ground for the
"HierarchUnitCubeDomainControl2". An upcoming commit will add a new
poisson benchmark app, which will take over this job.

Moreover, modified all applications to always output the number of
processes in the world communicator at app startup. Up to now, this
output was enclosed in an #ifdef, which was needed in the pre-FEAT::Dist
era.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>