minor revisement of tutorial applications
This commit introduces one 'major' change to all six tutorial codes:
The Geometry and LAFEM typedefs have been moved in front of the main
function, so that each of the tutorial namespaces starts with the
typedefs followed by the custom classes (e.g. AndicoreOperator) and
the actual main function.

This change should improve consistency with the more complex tutorials.
Moreover, separating the type definitions and the actual main function
should improve readability of the code.

Furthermore, some tutorials have been slightly modified so that the
ShapeType definition can be used to easily switch the whole application
to another shape type without the need to modify other code parts in
addition to the ShapeType typedef itself. However, some tutorials are
'hard-wired' to the quadrilateral shape and in this case a short
explaination was added.

Finally, this commit also includes lots of documentation fixes.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>