poisson_solver_factory app handles all shapes now
The poisson_solver_factory app has been slighly modified to support
tria/quad/tetra/hexa meshes by utilitising the fancy interplay of
Geometry::MeshFileReader and Control::Domain::PartiDomainControl :)

Moreover, the RHS of the PDE has been changed to =1, so now the PDE
reads

  -Laplace(u) = 1    in domain,
           u  = 0    on boundary.

As a result, the PDE is now senseful on any domai, i.e. this app can now
be applied on any mesh. As a drawback, the computation of errors against
a reference solution had to be removed, as there exists no known
reference solution anymore.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>