ENH: Nonlinear optimiser and line searches overhaul
The nonlinear optimisers (NLCG and NLSD) now both only store and use the
vanilla search direction, leaving it to the discretion of the line
search to normalise the search direction. The length of the vanilla
search direction can be used in the initial guess of the step length,
which can (and should!) only be used in conjunction with a
preconditioner by callding linesearch.set_dir_scaling(true).

This way, it was possible to move more shared functionality into the
base class.

Note that the MQCLinesearch has this feature disabled by default. Even
though it is theoretically 'better', the practical results are mixed at
best, and often downright worse.

Signed-off-by: Jordi Paul <jordi.paul@math.tu-dortmund.de>