implemented Tutorial 07: Stokes solver
As per Malte's request, a new tutorial demonstrating how to set up a
simple basic Stokes solver has been implemented and documented. Note
that (as usual), the tutorial is meant to be as simple as possible, thus
focusing on the main aspects like setting up the (non-global)
meta-containers and the saddle-point-matrix.

Additionally, this commit also introduces an implementation of
Assembly::Common::LaplaceOperatorBlocked as well as P0dc/P1dc element
template aliases, i.e. you can now use

  Space::Discontinuous::ElementP1<TrafoType>

instead of

  Space::Discontinuous::Element<TrafoType, Space::Discontinuous::Variant::StdPolyP<1>>

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>