use '--mesh-dir' param in Jordi's apps
Previously, Jordi's meshopt apps used the FEAT_SRC_DIR macro to
determine the path to the mesh files in the source code. As this
solution is not portable (and as Jordi is not here anymore), this
solution was replaced by a less convenient but more portable way,
namely by specifying the path to the directory that contains the
meshes by the '--mesh-path <dir>' parameters, just as it has been
done in the "navier_stokes_screws-app" for some time now.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>