rebranded 'SchurPrecond' to 'UzawaPrecond'
Unfortunately, the definition of 'Schur-Complement-Method' and 'Uzawa'
is quite mixed up in the literature and even more so in my head...

From my current understanding, our preconditioner implementation is
actually a simple 'inexact Uzawa' preconditioner, whereras most people
think of an iteration scheme including a virtual D*A^{-1}*B matrix when
reading/hearing about "Schur-Complement iteration/preconditioning".

Therefore, in the hope of clarification the 'SchurPrecond' has been
renamed to 'UzawaPrecond'.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>