added DistFileIO::write_ordered() function
This function is virtually a wrapper around 'MPI_File_write_ordered'
along with the necessary MPI I/O file open/close/truncate calls.

In layman's terms:
This function takes all the binary outputs of all processes and writes
them in rank order into one big common binary output file.

According to the MPI guys, this is the one and only way to go on large
clusters, as any attempt to create output files on a per-process level
is virtually equivalent to a DDoS attack on the file system...

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>