FEAT Geometry Module Fruehjahrsputz 2018
Dear Ladies and Gentlemen,
welcome to this year's episode of the FEAT Fruehjahrsputz!

This commit includes a lot of small changes in the mid- and low-level Geometry
code, as well as one change to the high-level code. Note that although most of
the changes are "under-the-hood" changes, and therefore most non-Geometry code
should remain unaffected, still a lot of files had to be modified mainly due to
two small (syntactic) interface changes.

The following changes have been implemented:

* Geometry::ConformalMesh
The "stride" template parameter has been removed from the ConformalMesh class
template, as it has never been used in a case where stride!=num_coords and
moreover many specialisations did not support that case anyway. Therefore,
removing this unwanted parameter seemed the right thing to do.
Note: This is one of the two interface changes that caused many syntactic (but
not semantic) file changes.

* Geometry::IndexSet
This class template has undergone several small changes:
- A new class template named 'IndexTuple' has been added to store a fixed-size
  tuple of indices (think of it as an std::array), which is now used instead of
  a fixed-size array typedef. This new class has become necessary because:
- The internal data array is now stored in a std::vector instead of a "naked"
  heap array. Due to this change, the IndexSet class is now copy- and movable.
- The IndexVectorType typedef has been replaced by the IndexTupleType typedef.
- The [Const]IndexVectorReference typedefs have been removed.
- The "operator()" now expects its second operand to be of type 'int' rather
  than 'Index', which improves consistency with the "local-indices-are-ints"
  rule that has been used throughout the FEAT kernel for quite some time now.
  Note: This is the second interface change that affected many syntactic
  (but not semantic) file changes, as operator this has required a lot of
  'int-to-Index' typecasts in the past, which now had to be removed.

* Geometry::VertexSet
This class has also undergone several small changes:
- The 'stride' template parameter has been removed for the same reasons as
  in the ConformalMesh class template.
- The vertices are now stored in an std::vector rather than a naked heap array.
  Due to this change, the VertexSet class is now copy- and movable.
- The [Const]VertexReference typedefs have been removed.

* Geometry::TargetSet
For the sake of consistency with the two previous class templates, the internal
data array is now stored in an std::vector, too, which also makes this class
copy- and movable.

* Geometry::MeshAttribute -> Geometry::AttributeSet
This class has undergone a set of severe changes:
- The MeshAttribute class has been renamed to AttributeSet to improve consistency
  with the naming of the other 'XyzSet' classes.
- The members of the new AttributeSet class have different names, e.g. 'values'
  instead of 'vertices'. This refactorisation has been long overdue as the
  'MeshAttribute' class was already a re-purposed version of the obsolete
  'VertexSetVariable' class.
- Again, the internal array is now handled by a std::vector, which also makes this
  class copy- and movable.

* Vertex / Attribute Refinement
As the vertex set and attribute set classes have now been fully decoupled, a new
StandardAttribRefiner class had to be written (although it is mostly identical
to the StandardVertexRefiner class). Moreover, as the VertexSet class has been
using Tiny::Vector objects to store the actual vertex coordinates for quite some
time now, which has rendered the auxiliary 'VertexAbacus' class obselete, so it
has been deleted.

Finally, two other classes have been renamed:
* PatchFactory --> PatchMeshFactory
* PatchHaloBuilder --> PatchHaloFactory

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>