added a checkpointing system
This commit enables a basic possibility to create checkpoints holding the state of all tracked objects during runtime.
Every checkpoint consists of two files:
the *.cp data file (*.zcp if compression with ZLib is used) and
the *.scp helper file (*.szcp if compression with ZLib is used)
Both files are needed to load a created checkpoint.

To use the checkpointing system create an instance of checkpoint_control and add the objects to be tracked using the add_object function.
A checkpoint containing all the tracked objects can now be created calling the save function.
The extension on the parameter filename specifies if the data should be written compressed (available only with ZLib support).
The load function extracts the data from the checkpoint and stores it within the checkpoint_control.
Obviously ZLib support is needed for loading compressed (*.zcp) files.
Then calling restore_object pushes back the previous saved state into the object.

Please note that loading a checkpoint is only possible with exactly the number of processes used to create it.

Signed-off-by: Kristina Sax <kristina.sax@math.tu-dortmund.de>