Supress nasty Wundef message in cuda headers
The compilers correctly warn on __CUDA_ARCH__ being not defined when
including the cuda_runtime.h header in host code.
Remember that the symbol must not be defined outside of device code.
As various versions of the cuda_runtime.h file have not been fixed and
the host+cuda_runtime situation only arises in one case, a simple hack
has been added.
The cuda header files are marked as system headers for the host compiler
by the -isystem flag und thus any warning concerning this files is
ignored (which is somewhat spooky, if you think about it a little
longer).

If someone knows a way to override the Wundef variants by a more
convenient way, independent of the compiler in use, he may be free to
replace this hack.

Signed-off-by: Dirk Ribbrock <dirk.ribbrock@math.tu-dortmund.de>