replaced nonlin/meshopt solver "print" functions by "info" functions
Many classes related to mesh-opt and to non-linear solvers featured a
member function called "print", which would dump some information about
the object to cout, usually using a Dist::Comm::world communicator.

This would have lead to serious problems aka deadlocks when a mesh-opt
tool and/or a non-linear solver was used on a non-world communicator,
e.g. on a sub-domain in a multi-physics simulation.

Therefore, the "print" functions have been replaced by an "info"
function, which simply returns a String object containing the information
about the object, which can then be pumped through a print call on
a corresponding communicator on application or control level.

Also removed a few redundant aka unused temproary world communicators
in some solver classes.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>