removed SolverBase::write_config() and its overrides
This function was the attempt to create an INI-file from a given solver tree,
which unfortunately made only sense in certain cases (e.g. no multigrid).

In the current code base, only the "dbg-nlopt" tool made use of this
functionality, and even there the tool had only a very limited set of
supported solver combinations, so actually creating the output by hand would
have been a much easier option here.

The provided functionality to dump a solver tree that was created from an
INI file back into an INI file is also quite pointless, unless one wants to
investigate whether this forms an idem-potent operation...

Finally, this functionality may become obsolete with the probably upcoming
"overhaul" of the solver factory anyways.

In consequence, the "dbg-nlopt" tool does not print the exact solver config
to the console anymore, but it still does its job aside from that.
If anybody thinks that this output was invaluable, (s)he is invited to dump
the selected solver configuration to cout directly in the tool code rather
than blowing up the whole solver interface for that. Thank you.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>