use Dist::Comm pointer in IterativeSolver base class
This commit adds a Dist::Comm* member to the IterativeSolver base class, which
is now used internally for the console output. This change has become necessary,
as the previous implementation was a mix of Comm::world-print calls and direct
std::cout outputs, which may have lead to problems with solvers on different
(nested) communicators.

Note that currently the top-level base class 'SolverBase' does not feature a
Dist::Comm pointer, as there is currently no need for that. Should the need
arise, we may think about moving the comm pointer directly into that base class.

This commit implies an important change:
Each solver that is derived from IterativeSolver (directly or indirectly) has
to specify the communicator of its system matrix/vectors upon construction,
preferably in the derived class's constructor.

This can be achieved in one of three possible ways:

1. Specify the communicator directly via the '_set_comm' member function.

2. Specify the communicator indirectly via the system matrix by using the
   '_set_comm_by_matrix' member function, which then extracts the communicator
   (if any) from the (global) system matrix.

3. Specify the communicator indirectly via a vector by using the
   '_set_comm_by_vector' member function, which then extracts the communicator
   (if any) from the (global) vector.

Currently, all solvers that derive from IterativeSolver receive a reference to
the system matrix upon construction, so that all current solvers use the second
approach, i.e. using the system matrix's communicator.

As a positive side-effect, the 'if(rank==0)' check around the 'set_plot_mode'
calls has become obsolete with this commit, as now the communicator takes care
that a parallel solver does not print its output multiple times. All applications,
tools and tutorials have been patched accordingly.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>