major analytic function overhaul and fixes
This commit fixes a major inconsistency with respect to the definition
of vector-valued function gradients and hessians that has been present
in the kernel for quite some time now.

Previously, the analytic function traits specified that the gradient
of a vector-valued function F: R^m --> R^n is a m-by-n matrix field
defined as follows:

(1)      grad[i][j] = d/d_i F_j

However, all vector fields in the Analytic::Common namespace as well
as the Assembly::VectorErrorComputer and Assembly::VelocityAnalyser
classes used the 'transposed' definition, i.e.

(2)      grad[i][j] = d/d_j F_i

Note that this inconsistency did not lead to any incorrect results,
because all related classes agreed to violate the definition (1)
and used definition (2) instead.

To resolve this inconsistency, the official definition of a vector-
valued function gradient has been changed from (1) to (2), i.e. now
the gradient of a vector field is the same as its jacobian matrix.
Note that the definition of the Hessian was also changed in the same
way, i.e. the hessian of a vector field F: R^m -> R^n is now a
m-by-n-by-n 3rd order tensor, where:

(3)      hess[i][j][k] = d/d_k d/d_j F_i

The decision for the change of these definitions was made with a few
other thoughts in mind:

First, now the definition of the gradient and the hessian of an analytic
function is the same as the definition of the jacobian and the hessian
of the transformation used by finite elment spaces. This should
hopefully reduce confusion when working with both of them.

Second, as matrices are typically stored and interpreted in row-major
order in C/C++, the new definitions (2) and (3) are favorable over the
old definitions, as many algorithms (e.g. the chain rule) can be written
in the sense of simple axpys, inner and outer as well as mat-vec products.
With the previous definitions, one would have required a lot of nested
loops that have somewhat weird order of indexing.

One possible drawback is that our new definition now breaks the common
(but not consent) "gradient = jacobian^T" rule.

Another possible (but currently uncertain) drawback is that the new
definition of gradients and hessians is now *possibly* inconsistent with
the definition used in Jordi's meshopt functionals, although it is
likely that this is not the case. Jordi has presumably used the "new"
definition in his code anyways, thus just aligning with all the other
old-definition-breakers. ^_^

Moreover, since I had to go through a lot of files for this commit to
ensure that I do not overlook anything, I have also used the opportunity
to slightly change the Analytic::Function interface:

Previously, the evaluation functions ('value, 'gradient', 'hessian') of
a function's Evaluator were expecting a reference to the output object
as a first function parameter. From this commit on, these functions now
return the result object, i.e. the signature of value evaluation function
is now

(4)      ValueType value(const PointType& point)

instead of (previously)

(5)      void value(ValueType& val, const PointType& point)

This should make working with analytic functions more convenient. The
original concerns about negative performance impacts have already been
eliminated by the compiler's return-value optimisation capabilities.
All analytic function implementations, wrapper and caller classes and
functions have been patched accordingly.

Finally, this commit also includes a few minor enhancements, which were
the original modification indent before noticing the above-mentioned
inconsistency that you have probably already forgotten about due to the
sheer length of this commit message ;)
These changes include, but are not limited to:

* Implemented gradient and hessian evaluation for Analytic::Common
  ::ParProfileScalar and ::ParProfileVector.

* Implemented H2-error computation in Assembly::VectorErrorComputer.

* Added H0/H1/H2 norms of the individual vector components to the
  Assembly::VectorErrorInfo class, which is returned as a result object
  by the Assembly::VectorErrorComputer.

* Added static member function 'null()' for Tiny::Vector, Tiny::Matrix
  and Tiny::Tensor3, which return a 'null' object of the corresponding
  class.

* Added Tiny::Tensor3::add_vec_mat_outer_product function(), which is
  used for the computation of the hessian tensor of a vector-valued
  finite element function.

* Fixed a few minor bugs in some Analytic::Common function gradient
  implementations.

* Removed a bunch of obsolete and erroneous copy-&-paste comments.

Thank you for our attention and have a nice and productive day :)

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>