implemented BurgersAssember::assemble_scalar_matrix
This function is the scalar "equivalent" of the vector-valued Burgers operator
assembler, which is nothing else than a Diffusion-Convection-Reaction operator
with a finite-element convection field.

Note that by construction there is no support for deformation tensors as well
as Frechet derivatives for scalar matrices, as this does not make sense.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>