implemented DistFileIO::read_ordered()
This commit adds the counterpart to the previously added write_ordered()
function. Note that in contrast to the other read_XXX() functions, this
function does not read the file contents into a stream, but directly into
a naked buffer. Also note that this function requires that the caller
knows how many bytes have to be read beforehand, as this function requires
this size as a parameter.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>