added VC14 support for HYPRE third-party lib
Note that the HYPRE sources have to be patched by using the provided 'vc_patch_hypre.py'
script as the library will not compile due to incorrect preprocessor defines otherwise.

Also renamed the old 'parmetis_patch_win.py' script to 'vc_patch_parmetis.py' for the
sake of consistency and updated the doxygen pages for third-party libs and the windows
build system.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>