added classes for Algebraic-DOF-Partitioning support
This commits adds three new major classes to the FEAT::Global namespace, which
can be used to "convert" a distributed linear system from its domain-
decomposition-based partitioning (used by the FEAT::Global classes) into an
(equivalent) algebraic partitioning based on the global DOFs, as well as a
solver-wrapper base-class outsourcing this dof-based partitioning.

The primary purpose of these classes is to enable us to use various parallel
third-party solver libraries, e.g. HYPRE, MUMPS, SuperLU and others, which
offer linear system solver frontends based on algebraic DOF partitioning.

The first and most important major class introduced by this commit is the
Global::AlgDofParti class, which computes an algebraic DOF partitioning from
a given Global::Gate object, which is then used by the AlgDofPartiMatrix/Vector
objects. The usage of this class is fire-&-forget, as all you have to do is
call its "assemble_by_gate()" function to do the dirty work for you.

Please note that the code will abort with a failed assertion if a process has
a patch with no interiour degrees of freedom *and* all neighbour processes
have a lower rank. Currently, there exists no solution to this problem, so just
pray that it doesn't happen -- or refine the mesh, if you can, because then
each patch will have at least one interiour degree of freedom.

Note that this is not a "bug" in the classical sense (it's not a "feature"
either), but circumventing this (hopefully rare) issue would require a
completely different (and more complex) partitioning approach, therefore the
"ostrich algorithm" was chosen here...

The second new class is the AlgDofPartiVector class, which represents a
partition of a global DOF vector and is therefore the algebraic-DOF-partitioned
counterpart of the domain-decomposition-based Global::Vector class. This class
offers two functions named "upload()" and "download()", which can be used to
copy the DOFs of a Global::Vector into an AlgDofPartiVector and vice versa.

The third new class is the AlgDofPartiMatrix class, which represents a row-
partitioned global CSR matrix. This class offers two functions named
"upload_symbolic()" and "upload_numeric(), which convert/copy the matrix layout
and values from a Global::Matrix object into the AlgDofPartiMatrix object.
Please note that (in contrast to vectors) it is *not* possible to "download"
a row-partitioned matrix into a Global::Matrix, because row-partitioned
matrices could contain matrix entries (i.e. dof couplings) which cannot be
represented in the domain-decomposition sense.

Please note that the AlgDofPartiMatrix and AlgDofPartiVector classes do *not*
implement the common LAFEM interface, i.e. there are no "axpy", "dot" or
"apply" member functions. In consequence, it is not possible to use these
classes in our "black-box" solvers directly or use them in any other manner
just as "usual" matrices and vectors. This is intentional and has two reasons:
First, some of these operations are non-trivial to implement and require
additional partitioning/communication magic and, second, this would then
compete with our preferred domain-decomposition approach. As already mentioned,
these classes are meant as "backend" classes for third-party libraries, which
then have to deal with the nasty communication issues on their own.

Furthermore, please note that the AlgDofParti classes can only work with scalar
systems, i.e. with the following domain-decomposition-based counterparts:
  * Global::Vector<LAFEM::DenseVector<Mem::Main, ...>, LAFEM::VectorMirror<Mem::Main,...>>
  * Global::Matrix<LAFEM::SparseMatrixCSR<Mem::Main, ...>, ...>

The reason for this is that the conversion between those two partitioning
strategies requires a lot of explicit DOF index transformations, which cannot
be out-sourced to the mirror classes in a black-box fashion.
This restriction may possibly be lifted in future.

Finally, the last new class is the Solver::ADPSolverBase class, which can be
used by third-party library solver wrapper classes to "out-source" the
conversion between domain-decomposition-based partitioning and algebraic DOF
partitioning into a common base-class. The primary purpose of this helper
class is to ensure that the derived classes do not have to deal with the
actual Global matrix and vector types and their conversion, which may be very
helpful when/if conversion for non-scalar systems is added in the future.

Signed-off-by: Peter Zajac <pzajac@math.tu-dortmund.de>